java_sdk_library: Remove some almost single use constants
The constants are only used in one place in the code and in the test.
This change switches the test code to use the apiScope methods to
generate the names and then inlines the constants in those methods. It
makes it easier to see the naming convention.

Test: m nothing
Bug: 155480189
Merged-In: I966e82b8ae67624da168e05f5a2b062b832cb8cd
Change-Id: I966e82b8ae67624da168e05f5a2b062b832cb8cd
(cherry picked from commit 542492e715a607c68fc3a619caaaec443c8c5adf)